Add camera control source, remove stale clients
Client heartbeats are now tracked via a Client class that
keeps track of the timestamp when the last heartbeat was
received. If this more time has elapsed than the
client_heartbeat_timeout_ms, the client is removed from
the clients dict. This allows a client to be disconnected
and reconnected without having to restart the server.